ipukms: fix ipu_get/clock_mutex interaction (lock debug)
Only ipu_get inside get_rate which is the only place it's truly required outside of
the IPU driver (and really even then, only for clock tree in debugfs). Otherwise the
pixel clock functionality is predicated on the caller having an ipu_get at some point
in it's past, so that the clock is enabled (ipu->usecount > 0) and these registers
are therefore accessible.